+++ ColourClues/Assets/Scenes/ExampleNetworkScene.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8c9cfa26abfee488c85f1582747f6a02
+guid: 2d4b6c8c253a0d346b4ba1a1db7b4428
 DefaultImporter:
   externalObjects: {}
   userData:
