+++ ColourClues/Assets/Scenes/Testing/ExampleNetworkScene.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d7bc506a8e7d0ae48a5ec30963b33571
+guid: 8c9cfa26abfee488c85f1582747f6a02
 DefaultImporter:
   externalObjects: {}
   userData:
