+++ ColourClues/Assets/Scripts/Player/Prefabs/NetworkTestingPlayer.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d7bc506a8e7d0ae48a5ec30963b33571
-DefaultImporter:
+guid: 7c8685e3f40ac1647a126ac32cb2c01d
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
