+++ ColourClues/Assets/Scripts/Player/Prefabs/Player.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5a8ff4deb8431294da24eadbd2134cc0
+guid: a8fa66eb0b413ee4cb846787717aafdb
 PrefabImporter:
   externalObjects: {}
   userData:
