+++ ColourClues/UserSettings/EditorUserSettings.asset
@@ -6,16 +6,16 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5b520d0503545b0d0c0c0a2715770748154f4d2c7d7d7e627a7d4a35b4e1646a
+      value: 5757035e52055b035b575a2143750a44434f4b2c7e7c77627e701966b1e46c68
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5b05065104065f590f5f0e274171064415151a73757a7765742c4931e0b56139
+      value: 5b520d0503545b0d0c0c0a2715770748154f4d2c7d7d7e627a7d4a35b4e1646a
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 0200575207005d5f580c0a7111715a4447164c297f7874332c7a4c61e0e46c3b
+      value: 5b05065104065f590f5f0e274171064415151a73757a7765742c4931e0b56139
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 5105045e00065e585d575d7248740b44144f1a2e7d2d20317d7e4f36bbb0376b
+      value: 07010c04010d5a03080c087b15755a44414e4e73787b24332f7d1962b0b26639
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
